java/de/uni/bamberg/wiai/minf/forflow/datagenerator/controller/datageneration/filltype/Country.java
package de.uni.bamberg.wiai.minf.forflow.datagenerator.controller.datageneration.filltype;

import java.util.ArrayList;
import java.util.Iterator;
import java.util.List;

import de.uni.bamberg.wiai.minf.forflow.datagenerator.controller.algorithm.ValueRetriever;

/**
 * This class has the job of generating country names.
 * The class inherits behaviour from its base-class,
 * that is {@link Generator}.
 * </p>
 * <font size=6><b>Observable</b></font></br>
 * It is observable like all other fill behaviour strategies.
 * Observable means in this context, whenever an observer 
 * registers to or un-registers from this class, the subject
 * itself notifies the UI to register this class in a specific
 * list of generators.
 * 
 * @author Michael Munz
 * @version 0.1
 * @since May/11/09
 */
public class Country extends Generator
{
	/**
	 * holds the list of observers of this class
	 */
	private List<ObserverGenerator> observers = null;
	
	/**
	 * default constructor
	 */
	public Country()
	{
		super("Country");
		
		this.setPathToFileByLanguage();
	}
	
	protected Country(Country country)
	{
		super(country);
	}
	
	@Override
	public String getValueAt(int n)
	{
		this.setPathToFileByLanguage();
		
		// is in interval?
		if((n >= 0) && (n < this.getN()))
		{
			return ValueRetriever.getValueAt(n, getPathToFile());
		}
		else
		{
			return null;
		}
	}
	
	@Override
	public void registerObserverGenerator(ObserverGenerator observer)
	{
		if(this.observers == null)
		{
			this.observers = new ArrayList<ObserverGenerator>();
		}
		
		this.observers.add(observer);
		
		this.notifyObserverGenerator(true);
	}
	
	@Override
	public void removeObserverGenerator(ObserverGenerator observer)
	{
		if(this.observers != null)
		{
			int n = this.observers.indexOf(observer);
			
			if(n >= 0)
			{
				this.observers.remove(n);
				
				this.notifyObserverGenerator(false);
			}
		}
	}
	
	@Override
	public void notifyObserverGenerator(boolean add)
	{
		for(Iterator<ObserverGenerator> i=this.observers.iterator(); i.hasNext();)
		{
			ObserverGenerator ob = i.next();
			
			ob.updateGenerator(add, this);
		}
	}
	
	@Override
	public Object clone() throws CloneNotSupportedException
	{
		return new Country(this);
	}
	
	@Override
	protected String getFile()
	{
		return Resources.COUNTRIES;
	}
}